Extract language section to component.
